# docs/swagger.yaml
basePath: /
definitions:
...
  internal_modules_products.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  internal_modules_products.CategoryListSuccessResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_modules_products.Category'
        type: array
      errors:
        type: object
      message:
        example: Categories
        type: string
      success:
        example: true
        type: boolean
    type: object
  internal_modules_products.CategoryResponse:
    properties:
      category:
        $ref: '#/definitions/internal_modules_products.Category'
    type: object
  internal_modules_products.CategorySuccessResponse:
    properties:
      data:
        $ref: '#/definitions/internal_modules_products.CategoryResponse'
      errors:
        type: object
      message:
        example: Category
        type: string
      success:
        example: true
        type: boolean
    type: object
  internal_modules_products.CreateCategoryRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 120
        minLength: 2
        type: string
      slug:
        maxLength: 140
        minLength: 2
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - name
    - slug
    type: object
  internal_modules_products.CreateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      height_mm:
        type: integer
      length_mm:
        type: integer
      main_image_url:
        maxLength: 500
        type: string
      material:
        maxLength: 120
        type: string
      name:
        maxLength: 180
        minLength: 2
        type: string
      price:
        minimum: 0
        type: number
      short_description:
        maxLength: 300
        type: string
      sku:
        maxLength: 80
        type: string
      slug:
        maxLength: 200
        minLength: 2
        type: string
      status:
        enum:
        - draft
        - active
        - inactive
        type: string
      stock:
        type: integer
      weight_gram:
        type: integer
      width_mm:
        type: integer
    required:
    - name
    - price
    - slug
    type: object
  internal_modules_products.ErrorResponse:
    properties:
      data:
        type: object
      errors:
        type: object
      message:
        example: Validation failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  internal_modules_products.ListProductsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_modules_products.Product'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  internal_modules_products.MessageResponse:
    properties:
      data:
        type: object
      errors:
        type: object
      message:
        example: Product deleted successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  internal_modules_products.Product:
    properties:
      category:
        $ref: '#/definitions/internal_modules_products.Category'
      category_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      height_mm:
        type: integer
      id:
        type: integer
      length_mm:
        type: integer
      main_image_url:
        type: string
      material:
        type: string
      name:
        type: string
      price:
        type: number
      short_description:
        type: string
      sku:
        type: string
      slug:
        type: string
      status:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: integer
      weight_gram:
        type: integer
      width_mm:
        type: integer
    type: object
  internal_modules_products.ProductListSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/internal_modules_products.ListProductsResponse'
      errors:
        type: object
      message:
        example: Products
        type: string
      success:
        example: true
        type: boolean
    type: object
  internal_modules_products.ProductResponse:
    properties:
      product:
        $ref: '#/definitions/internal_modules_products.Product'
    type: object
  internal_modules_products.ProductSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/internal_modules_products.ProductResponse'
      errors:
        type: object
      message:
        example: Product
        type: string
      success:
        example: true
        type: boolean
    type: object
  internal_modules_products.UpdateCategoryRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 120
        minLength: 2
        type: string
      slug:
        maxLength: 140
        minLength: 2
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - name
    - slug
    type: object
  internal_modules_products.UpdateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      height_mm:
        type: integer
      length_mm:
        type: integer
      main_image_url:
        maxLength: 500
        type: string
      material:
        maxLength: 120
        type: string
      name:
        maxLength: 180
        minLength: 2
        type: string
      price:
        minimum: 0
        type: number
      short_description:
        maxLength: 300
        type: string
      sku:
        maxLength: 80
        type: string
      slug:
        maxLength: 200
        minLength: 2
        type: string
      status:
        enum:
        - draft
        - active
        - inactive
        type: string
      stock:
        type: integer
      weight_gram:
        type: integer
      width_mm:
        type: integer
    required:
    - name
    - price
    - slug
    type: object
info:
  contact: {}
  description: API documentation for EDNIC backend.
  title: EDNIC Backend API
  version: 0.1.0
paths:
...
  /api/v1/admin/product-categories:
    get:
      description: Returns all product categories including inactive categories.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.CategoryListSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List product categories for admin
      tags:
      - Admin Products
    post:
      consumes:
      - application/json
      description: Creates a product category from admin panel.
      parameters:
      - description: Create category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_products.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.CategorySuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create product category
      tags:
      - Admin Products
  /api/v1/admin/product-categories/{id}:
    delete:
      description: Deletes a product category from admin panel.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete product category
      tags:
      - Admin Products
    put:
      consumes:
      - application/json
      description: Updates a product category from admin panel.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_products.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.CategorySuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product category
      tags:
      - Admin Products
  /api/v1/admin/products:
    get:
      description: Returns products for admin including draft and inactive items.
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Category slug
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.ProductListSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List products for admin
      tags:
      - Admin Products
    post:
      consumes:
      - application/json
      description: Creates a product from admin panel.
      parameters:
      - description: Create product request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_products.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.ProductSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - Admin Products
  /api/v1/admin/products/{id}:
    delete:
      description: Deletes product from admin panel.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Admin Products
    get:
      description: Returns product detail by id for admin.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.ProductSuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get product for admin
      tags:
      - Admin Products
    put:
      consumes:
      - application/json
      description: Updates product from admin panel.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update product request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_products.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.ProductSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Admin Products
...
  /api/v1/product-categories:
    get:
      description: Returns active product categories.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.CategoryListSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      summary: List public product categories
      tags:
      - Products
  /api/v1/products:
    get:
      description: Returns active products for public catalog.
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Category slug
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.ProductListSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      summary: List public products
      tags:
      - Products
  /api/v1/products/{slug}:
    get:
      description: Returns active product detail by slug.
      parameters:
      - description: Product slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_products.ProductSuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_modules_products.ErrorResponse'
      summary: Get public product detail
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: 'Use format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"